refactor: split betterAuth.schema into individual schema files (session, verification, account)
